Allow use of 'notes open' when no TTY is available
This is useful for a) our automated tests but also b) any
scripting of notes. You probably don't want to open an
editor in your terminal in that case, but you could
plausibly end up running 'notes open' at some point anyway
and it's nice if it doesn't crash at least.