Fix /dev/tty check
Previously this was bugged, incorrectly using -z so that the check never
passed, and $EDITOR was always run without having the TTY injected as
an input. This causes problems with command-line editors like Vim, which
need an interactive input (and output) to be usable.

Meanwhile, attempting to inject /dev/tty as the input in all cases
causes problems in CI, because there is no TTY available in GitHub
Actions, and presumably in other scripted environments.

Fixing this with -e does not work, as /dev/tty exists but unopenably in
many environments. Instead, we use printf in a subshell to test whether
/dev/tty is actually usable, and then only use it if it is.